modified the codes to be consistent with latest and greatest versions of the dependencies.

+++ popgen/draw.py
@@ -86,11 +86,11 @@ class Draw_Population(object):
 
     def _pick_households(self, geo_id_frequencies, geo_cumulative_weights):
         last = 0
-        rand_numbers = np.random.random(geo_id_frequencies.sum())
+        rand_numbers = np.random.random(geo_id_frequencies.sum().astype(int))
         list_rows_syn_subpop = []
         for column in self.geo_frequencies.columns.values:
             rows = self.geo_row_idx[column]
-            column_frequency = geo_id_frequencies[column]
+            column_frequency = geo_id_frequencies[column].astype(int)
             column_bins = np.searchsorted(geo_cumulative_weights[column],
                                           (rand_numbers[
                                            last:last + column_frequency]),
